+++ src/main/java/com/michelet/restaurant/MicheletApplication.java
@@ -1,4 +1,4 @@
-package com.star.michelet;
+package com.michelet.restaurant;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
