+++ src/test/java/com/michelet/restaurant/MicheletApplicationTests.java
@@ -1,4 +1,4 @@
-package com.star.michelet;
+package com.michelet.restaurant;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
